requirements.txt
django
channels
asgi_redis
feedparser